Change type of GetEntryInfoResult to int32 to match type of FhgfsOpsErr

+++ common/ioctl/beegfs.go
@@ -239,5 +239,6 @@ type getEntryInfoV2Arg struct {
 	FileDataState    uint8
 	// Result of the GetEntryInfo RPC. 0 = success (full data populated). Non-zero = partial
 	// result (only basic entry info valid); value is the FhgfsOpsErr code from the metadata RPC.
-	GetEntryInfoResult uint8
+	// Matches the signed 32-bit wire type of FhgfsOpsErr.
+	GetEntryInfoResult int32
 }
